Updated Projects, Added Favicon, and Other Minor Changes
Haven't updated the old homepage in a while. Maybe I'll get around to a
facelift one of these days. For now, I've added the some of the larger
projects I've done in the past couple years.

I also added a favicon, renamed Blog to The Feed, and fixed project images.